feat(svelte): add Tiptap context provider with getEditor() and Tiptap.Content
Adds a <Tiptap editor={editor}> provider component using Svelte 5 setContext/getContext, a <Tiptap.Content /> sub-component that reads the editor from context, and a getEditor() helper for child components to access the editor without prop drilling.

- editorContext.ts: shared Symbol key + setEditor/getEditor helpers
- Tiptap.svelte: provider that calls setEditor() and renders children
- TiptapContent.svelte: calls getEditor() and delegates to EditorContent
- index.ts: Tiptap = Object.assign(TiptapRoot, { Content: TiptapContent })
- demos: Default/Svelte uses Tiptap + Toolbar (via getEditor), Menus/Svelte uses Tiptap + Tiptap.Content
