+++ packages/svelte/src/components/TiptapContent.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+  import { getEditor } from './editorContext.js'
+  import EditorContent from './EditorContent.svelte'
+
+  const editor = getEditor()
+</script>
+
+<EditorContent editor={editor} />
